Finished implementing Drawables
types that implement Drawable interface can be added to a layer and drawn by engine.Drawer

TODO: make it read from an enum to add layers

Made the Drawer be the default event listener. Might move this to a different class that can handle multiple engines.

Main now invokes the Renderer.init() and the desired engines.